+++ pkg/bunapp/chmigrations/20220703121853_metrics.up.sql
@@ -8,7 +8,7 @@ CREATE TABLE ?DB.measure_minutes ?ON_CLUSTER (
   sum SimpleAggregateFunction(sum, Float64) Codec(?CODEC),
   count SimpleAggregateFunction(sum, UInt64) Codec(T64, ?CODEC),
   value SimpleAggregateFunction(anyLast, Float64) Codec(?CODEC),
-  histogram AggregateFunction(quantilesBFloat16(0.5, 0.9, 0.99), Float32) Codec(?CODEC),
+  histogram AggregateFunction(quantilesBFloat16(0.5), Float32) Codec(?CODEC),
 
   attr_keys Array(LowCardinality(String)) Codec(?CODEC),
   attr_values Array(LowCardinality(String)) Codec(?CODEC)
@@ -37,7 +37,7 @@ CREATE TABLE ?DB.measure_hours ?ON_CLUSTER (
   sum SimpleAggregateFunction(sum, Float64) Codec(?CODEC),
   count SimpleAggregateFunction(sum, UInt64) Codec(T64, ?CODEC),
   value SimpleAggregateFunction(anyLast, Float64) Codec(?CODEC),
-  histogram AggregateFunction(quantilesBFloat16(0.5, 0.9, 0.99), Float32) Codec(?CODEC),
+  histogram AggregateFunction(quantilesBFloat16(0.5), Float32) Codec(?CODEC),
 
   attr_keys Array(LowCardinality(String)) Codec(?CODEC),
   attr_values Array(LowCardinality(String)) Codec(?CODEC)
@@ -63,31 +63,10 @@ AS SELECT
   sum(sum) AS sum,
   sum(count) AS count,
   anyLast(value) AS value,
-  quantilesBFloat16MergeState(0.5, 0.9, 0.99)(histogram) AS histogram,
+  quantilesBFloat16MergeState(0.5)(histogram) AS histogram,
 
   anyLast(attr_keys) AS attr_keys,
   anyLast(attr_values) AS attr_values
 FROM ?DB.measure_minutes
 GROUP BY project_id, metric, toStartOfHour(time), attrs_hash
-SETTINGS prefer_column_name_to_alias = 1
-
---migration:split
-
-CREATE MATERIALIZED VIEW ?DB.spans_metrics_mv ?ON_CLUSTER
-TO ?DB.measure_minutes AS
-SELECT
-  project_id,
-  'uptrace.spans.duration' AS metric,
-  toStartOfMinute(time) AS time,
-  xxHash64(arrayStringConcat([system, host_name, service_name], '-')) AS attrs_hash,
-
-  'histogram' AS instrument,
-  sum(duration) AS sum,
-  count() AS count,
-  quantilesBFloat16StateIf(0.5, 0.9, 0.99)(toFloat32(duration / 1000), duration > 0) AS histogram,
-
-  ['span.system', 'host.name', 'service.name'] AS attr_keys,
-  [system, host_name, service_name] AS attr_values
-FROM ?DB.spans_index
-GROUP BY project_id, toStartOfMinute(time), system, host_name, service_name
 SETTINGS prefer_column_name_to_alias = 1