+++ pkg/metrics/const.go
@@ -1,0 +1,10 @@
+package metrics
+
+const (
+	InvalidInstrument   = ""
+	GaugeInstrument     = "gauge"
+	AdditiveInstrument  = "additive"
+	HistogramInstrument = "histogram"
+	CounterInstrument   = "counter"
+	SummaryInstrument   = "summary"
+)